# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI Creator Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /payments:
    get:
      operationId: listPayments
      tags: [payments]
      summary: List payments
      parameters:
        - in: query
          name: campaignId
          schema:
            type: integer
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, escrow, released, refunded]
      responses:
        "200":
          description: Payments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"
    post:
      operationId: createPayment
      tags: [payments]
      summary: Create a payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaymentBody"
      responses:
        "201":
          description: Payment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
components:
  schemas:
...
    Payment:
      type: object
      properties:
        id:
          type: integer
        campaignId:
          type: integer
        campaignTitle:
          type: string
        brandId:
          type: integer
        creatorId:
          type: integer
        creatorName:
          type: string
        amount:
          type: number
        status:
          type: string
          enum: [pending, escrow, released, refunded]
        platformFee:
          type: number
        creatorPayout:
          type: number
        createdAt:
          type: string
      required:
        - id
        - campaignId
        - brandId
        - creatorId
        - amount
        - status
        - createdAt
    CreatePaymentBody:
      type: object
      properties:
        campaignId:
          type: integer
        brandId:
          type: integer
        creatorId:
          type: integer
        amount:
          type: number
      required:
        - campaignId
        - brandId
        - creatorId
        - amount